+++ src/OperationBar/OperationSelection.tsx
@@ -10,16 +10,24 @@ import { Icon } from "../components/Icon";
 import { IconButton } from "../components/IconButton";
 import { editingModeState, EditingModes } from "../Store";
 import { useRecoilState } from "recoil";
+import { useContext } from "react";
+import {
+  AdjustmentOperations,
+  EditingOperations,
+  EditorContext,
+  TransformOperations,
+} from "../ImageEditor";
+import { useMemo } from "react";
 
-interface Operation {
+interface Operation<T> {
   title: string;
   iconID: string;
-  operationID: EditingModes;
+  operationID: T;
 }
 
 interface Operations {
-  transform: Operation[];
-  adjust: Operation[];
+  transform: Operation<TransformOperations>[];
+  adjust: Operation<AdjustmentOperations>[];
 }
 
 const operations: Operations = {
@@ -46,45 +54,92 @@ const operations: Operations = {
 
 export function OperationSelection() {
   //
-  const [selectedOperation, setSelectedOperation] = React.useState<
+  const {
+    allowedTransformOperations,
+    allowedAdjustmentOperations,
+  } = useContext(EditorContext);
+
+  const isTransformOnly =
+    allowedTransformOperations && !allowedAdjustmentOperations;
+  const isAdjustmentOnly =
+    allowedAdjustmentOperations && !allowedTransformOperations;
+
+  const [selectedOperationGroup, setSelectedOperationGroup] = React.useState<
     "transform" | "adjust"
-  >("transform");
+  >(isAdjustmentOnly ? "adjust" : "transform");
 
   const [, setEditingMode] = useRecoilState(editingModeState);
+
+  const filteredOperations = useMemo(() => {
+    // If neither are specified then allow the full range of operations
+    if (!allowedTransformOperations && !allowedAdjustmentOperations) {
+      return operations;
+    }
+    const filteredTransforms = allowedTransformOperations
+      ? operations.transform.filter((op) =>
+          allowedTransformOperations.includes(op.operationID)
+        )
+      : operations.transform;
+    const filteredAdjustments = allowedAdjustmentOperations
+      ? operations.adjust.filter((op) =>
+          allowedAdjustmentOperations.includes(op.operationID)
+        )
+      : operations.adjust;
+    if (isTransformOnly) {
+      return { transform: filteredTransforms, adjust: [] };
+    }
+    if (isAdjustmentOnly) {
+      return { adjust: filteredAdjustments, transform: [] };
+    }
+    return { transform: filteredTransforms, adjust: filteredAdjustments };
+  }, [
+    allowedTransformOperations,
+    allowedAdjustmentOperations,
+    isTransformOnly,
+    isAdjustmentOnly,
+  ]);
 
   return (
     <>
       <ScrollView style={styles.opRow} horizontal>
-        {operations[selectedOperation].map((item, index) => (
-          <View style={styles.opContainer} key={item.title}>
-            <IconButton
-              text={item.title}
-              iconID={item.iconID}
-              onPress={() => setEditingMode(item.operationID)}
-            />
-          </View>
-        ))}
+        {filteredOperations[selectedOperationGroup].map(
+          (item: Operation<EditingOperations>, index: number) => (
+            <View style={styles.opContainer} key={item.title}>
+              <IconButton
+                text={item.title}
+                iconID={item.iconID}
+                onPress={() => setEditingMode(item.operationID)}
+              />
+            </View>
+          )
+        )}
       </ScrollView>
-      <View style={styles.modeRow}>
-        <TouchableOpacity
-          style={[
-            styles.modeButton,
-            selectedOperation === "transform" && { backgroundColor: "#333" },
-          ]}
-          onPress={() => setSelectedOperation("transform")}
-        >
-          <Icon iconID="transform" text="Transform" />
-        </TouchableOpacity>
-        <TouchableOpacity
-          style={[
-            styles.modeButton,
-            selectedOperation === "adjust" && { backgroundColor: "#333" },
-          ]}
-          onPress={() => setSelectedOperation("adjust")}
-        >
-          <Icon iconID="tune" text="Adjust" />
-        </TouchableOpacity>
-      </View>
+      {!isTransformOnly && !isAdjustmentOnly ? (
+        <View style={styles.modeRow}>
+          <TouchableOpacity
+            style={[
+              styles.modeButton,
+              selectedOperationGroup === "transform" && {
+                backgroundColor: "#333",
+              },
+            ]}
+            onPress={() => setSelectedOperationGroup("transform")}
+          >
+            <Icon iconID="transform" text="Transform" />
+          </TouchableOpacity>
+          <TouchableOpacity
+            style={[
+              styles.modeButton,
+              selectedOperationGroup === "adjust" && {
+                backgroundColor: "#333",
+              },
+            ]}
+            onPress={() => setSelectedOperationGroup("adjust")}
+          >
+            <Icon iconID="tune" text="Adjust" />
+          </TouchableOpacity>
+        </View>
+      ) : null}
     </>
   );
 }