+++ src/Store.tsx
@@ -1,7 +1,8 @@
 import { ExpoWebGLRenderingContext } from "expo-gl";
+import { EditingOperations } from "expo-image-editor";
 import { atom } from "recoil";
 
-interface ImageData {
+export interface ImageData {
   uri: string;
   height: number;
   width: number;
@@ -10,7 +11,7 @@ interface ImageData {
 export const imageDataState = atom<ImageData>({
   key: "imageDataState",
   default: {
-    uri: undefined,
+    uri: "",
     width: 0,
     height: 0,
   },
@@ -21,7 +22,7 @@ export const imageScaleFactorState = atom<number>({
   default: 1,
 });
 
-interface ImageBounds {
+export interface ImageBounds {
   x: number;
   y: number;
   height: number;
@@ -48,7 +49,7 @@ export const processingState = atom<boolean>({
   default: false,
 });
 
-interface AccumulatedPan {
+export interface AccumulatedPan {
   x: number;
   y: number;
 }
@@ -61,7 +62,7 @@ export const accumulatedPanState = atom<AccumulatedPan>({
   },
 });
 
-interface ImageDimensions {
+export interface ImageDimensions {
   width: number;
   height: number;
 }
@@ -74,29 +75,11 @@ export const cropSizeState = atom<ImageDimensions>({
   },
 });
 
-export type EditingModes = "operation-select" | "crop" | "rotate" | "blur";
+export type EditingModes = "operation-select" | EditingOperations;
 
 export const editingModeState = atom<EditingModes>({
   key: "editingModeState",
   default: "operation-select",
-});
-
-export const fixedCropAspectRatioState = atom<number>({
-  key: "fixedCropAspectRatioState",
-  default: 1,
-});
-
-export const lockAspectRatioState = atom<boolean>({
-  key: "lockAspectRatioState",
-  default: false,
-});
-
-export const minimumCropDimensionsState = atom<ImageDimensions>({
-  key: "minimumCropDimensionsState",
-  default: {
-    width: 0,
-    height: 0,
-  },
 });
 
 interface GLContext {
@@ -110,12 +93,7 @@ export const glContextState = atom<GLContext["gl"]>({
   default: null,
 });
 
-export const glProgramState = atom<GLContext["program"]>({
+export const glProgramState = atom<GLContext["program"] | null>({
   key: "glProgramState",
   default: null,
-});
-
-export const throttleBlurState = atom<boolean>({
-  key: "throttleBlurState",
-  default: false,
 });