+++ CHANGELOG.md
@@ -2,4 +2,21 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.5.0](https://github.com/thomas-coldwell/expo-image-editor/compare/v1.4.4...v1.5.0) (2021-05-06)
+
+
+### Features
+
+* Ability to specify `allowedTransformOperations` and `allowedAdjustmentOperations` to control what operations the use has access to in full mode ([9a7224d](https://github.com/thomas-coldwell/expo-image-editor/commit/9a7224d4df3b67fb485532f2cef6212f933e5ad5))
+* Crop only mode with custom hook to make cropping functionaility reusable ([798e9b0](https://github.com/thomas-coldwell/expo-image-editor/commit/798e9b02d9ccfce4af216233a5a0db6d26529b10))
+* Improved layout of the example ([6d52dee](https://github.com/thomas-coldwell/expo-image-editor/commit/6d52dee99eb4de2e5c922e3aed2b76bd5501b703))
+* Updated RNGH deps ([e2daf03](https://github.com/thomas-coldwell/expo-image-editor/commit/e2daf031975bacac51dd7bef4f640e90557123db))
+
+
+### Bug Fixes
+
+* Added Standard Version ([cbd2bb7](https://github.com/thomas-coldwell/expo-image-editor/commit/cbd2bb7d0bfdc797946f0230545110de3f078afd))
+* Image crop overlay window not regsitering touches on Android in modal ([6aeb120](https://github.com/thomas-coldwell/expo-image-editor/commit/6aeb1201923ad00116e3672f8834d6042a4d66be))
+* Missing types and typescript errors ([5fbdad6](https://github.com/thomas-coldwell/expo-image-editor/commit/5fbdad6948460e2ddfa988ca73cfa93d77d8d1bf))
+
 ### 1.4.4 (2021-04-23)